pyproject.toml
[project]
name = "planetarium-api-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "flake8 (>=7.2.0,<8.0.0)",
    "flake8-quotes (>=3.4.0,<4.0.0)",
    "flake8-variables-names (>=0.0.6,<0.0.7)",
    "pep8-naming (>=0.15.1,<0.16.0)",
    "django-debug-toolbar (>=5.2.0,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
